chore(filter): minor tweaks to filter package.
