+++ sdk/filter/README.md
@@ -9,10 +9,12 @@ Example
 type FooURLFilter struct {
 }
 
+// Filter evaluates whether request should be blocked, `true` blocks the request and `false` continues it.
 func (FooURLFilter) EvaluateURLAndHeaders(span sdk.Span, url string, headers map[string][]string) bool {
 	return false
 }
 
+// Filter evaluates whether request should be blocked, `true` blocks the request and `false` continues it.
 func (FooURLFilter) EvaluateBody(span sdk.Span, body []byte) bool {
 	return false
 }
