+++ sdk/filter/noop_test.go
@@ -1,0 +1,13 @@
+package filter
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestNoopFilter(t *testing.T) {
+	f := NoopFilter{}
+	assert.False(t, f.EvaluateURLAndHeaders(nil, "", nil))
+	assert.False(t, f.EvaluateBody(nil, nil))
+}
